pyproject.toml
[tool.poetry]
name = "armada-security"
version = "0.5.0"
description = "Armada Security Tools"
authors = ["OmniVector Solutions <info@omnivector.solutions>"]
license = "MIT"
packages = [ { include = "armasec" } ]

[tool.poetry.dependencies]
python = "^3.8"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
fastapi = "^0.68.0"
httpx = "^0.18.2"
snick = "^1.0.0"
py-buzz = "^2.1.3"

[tool.poetry.dev-dependencies]
ipython = "^7.26.0"
pytest-freezegun = "^0.4.2"
asgi-lifespan = "^1.0.1"
httpx = "^0.18.2"
pytest-asyncio = "^0.15.1"
pytest-random-order = "^1.0.4"
mypy = "^0.910"
isort = "^5.9.3"
pytest-sugar = "^0.9.4"
black = "^21.7b0"
flake8 = "^3.9.2"
respx = "^0.17.1"
python-dotenv = "^0.19.0"
snick = "^1.0.0"
pytest-cov = "^2.12.1"

[tool.poetry.plugins.pytest11]
json_decoder = "armasec.pytest_extension"

[tool.pytest.ini_options]
minversion = "6.0"
# Unfortunately, pytest-cov breaks when we add armasec.pytest_extension as a pytest plugin
# TODO: Figure out why this is and how to fix it
# addopts = "--random-order --cov=armasec --cov-report=term-missing --cov-fail-under=85"
addopts = "--random-order"
testpaths = ["tests"]

[[tool.mypy.overrides]]
module = [
    "jose",
    "buzz",
    "snick",
]
ignore_missing_imports = true

[tool.black]
line-length = 100

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
